mqqt-py/requirements.txt
paho-mqtt==1.6.1

sqlalchemy

pytest

mysql-connector-python